Complete functionality of the signup page and prepare the signupDetails object to be used in the future
