+++ package.json
@@ -20,6 +20,7 @@
         "axios": "^0.21.1",
         "axios-auth-refresh": "^3.1.0",
         "date-fns": "^2.20.2",
+        "form-data": "^4.0.0",
         "lodash": "^4.17.21",
         "moment": "^2.29.1",
         "node-sass": "^5.0.0",
